fix: style upload dialog
